requirements.txt
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
matplotlib==3.10.1
networkx==3.4.2
numpy==2.2.3
pandas==2.2.3
pillow==11.1.0
pydot==3.0.4
pygraphviz==1.14
PyYAML==6.0.2
tabulate==0.9.0
colorama==0.4.6